fixed make.sh

+++ make.sh
@@ -1,8 +1,8 @@
 #!/bin/sh
 
-rm ./build/source
+[[ -e build/source ]] && rm ./build/source
 echo "compiling libsrc"
-g++ src/libsrc/*.cc -I include -fPIC -shared -o build/libsrc -std=c++17
+g++ src/libsrc/*.cc -I include -fPIC -shared -o build/libsrc.so -std=c++17 -fpermissive -Wno-int-to-pointer-cast -Wno-return-type
 
 echo "compiling source"
 g++ src/source/*.cc -I include -o build/source -std=c++17 -Lbuild/ -lsrc
